+++ database/migrations/2021_02_02_091330_create_stock_barang_bulkies_table.php
@@ -18,7 +18,7 @@ class CreateStockBarangBulkiesTable extends Migration
             $table->foreignId('bulky_id')->constrained('gudang_bulkies')->onDelete('cascade');
             $table->string('barang_kode');
             $table->foreign('barang_kode')->references('kode_barang')->on('barangs')->onDelete('cascade');
-            $table->bigInteger('jumlah');
+            $table->float('jumlah', 11, 2);
             $table->string('satuan');
             $table->bigInteger('harga_barang')->nullable();
             $table->float('diskon', 11, 2)->nullable();
